use float32

+++ drGAT/drGAT.py
@@ -115,6 +115,7 @@ class drGAT(Module):
         all_attention = get_attention_mat(attention)
         del attention
 
+        x = x.to(torch.float32)
         x = self.dropout1(self.activation(self.graph_norm1(x)))
 
         x, attention = self.gat2(
@@ -125,7 +126,8 @@ class drGAT(Module):
         )
         all_attention += get_attention_mat(attention)
         del attention
-
+        
+        x = x.to(torch.float32)
         x = self.dropout2(self.activation(self.graph_norm2(x)))
 
         x = torch.concat(
